+++ webs/sync-web/docroot/WEB-INF/service/com/liferay/sync/service/SyncDeviceServiceUtil.java
@@ -64,7 +64,7 @@ public class SyncDeviceServiceUtil {
 	}
 
 	public static com.liferay.sync.model.SyncDevice registerSyncDevice(
-		java.lang.String type, int buildNumber, int featureSet,
+		java.lang.String type, long buildNumber, int featureSet,
 		java.lang.String uuid)
 		throws com.liferay.portal.kernel.exception.PortalException,
 			com.liferay.portal.kernel.exception.SystemException {
