+++ webs/sync-web/docroot/WEB-INF/service/com/liferay/sync/service/SyncDeviceServiceWrapper.java
@@ -58,7 +58,7 @@ public class SyncDeviceServiceWrapper implements SyncDeviceService,
 
 	@Override
 	public com.liferay.sync.model.SyncDevice registerSyncDevice(
-		java.lang.String type, int buildNumber, int featureSet,
+		java.lang.String type, long buildNumber, int featureSet,
 		java.lang.String uuid)
 		throws com.liferay.portal.kernel.exception.PortalException,
 			com.liferay.portal.kernel.exception.SystemException {
